+++ plugins/Shortcodes.php
@@ -28,7 +28,7 @@ class Shortcodes {
 		\add_shortcode('four_columns_two', array($this, 'shortcodeFourColumnsTwo'));
 		\add_shortcode('four_columns_three', array($this, 'shortcodeFourColumnsThree'));
 		\add_shortcode('divider', array($this, 'shortcodeDivider'));
-		\add_shortcode('button', array($this, 'shortcodeButton'));
+//		\add_shortcode('button', array($this, 'shortcodeButton'));
 		\add_shortcode('credits', array($this, 'shortcodeCredits'));
 	} // END public function registerShortcodes()
 
@@ -208,26 +208,26 @@ class Shortcodes {
 		return '<div class="divider clearfix"></div>';
 	} // END public function shortcodeDivider($atts, $content = null)
 
-	public function shortcodeButton($atts, $content = null) {
-		$args = \shortcode_atts(
-			array(
-				'link' => '',
-				'type' => 'default',
-				'target' => '',
-				'size' => ''
-			),
-			$atts
-		);
-
-		$type = (!empty($args['type'])) ? ' btn-' . $args['type'] : '';
-		$link = $args['link'];
-		$target = $args['target'];
-		$size = (!empty($args['size'])) ? ' btn-' . $args['size'] : '';
-
-		$output = '<a class="btn ' . $type . $size . '" href="' . $link . '" target="' . $target . '"><span>' . \do_shortcode($content) . '</span></a>';
-
-		return $output;
-	} // END public function shortcodeButton($atts, $content = null)
+//	public function shortcodeButton($atts, $content = null) {
+//		$args = \shortcode_atts(
+//			array(
+//				'link' => '',
+//				'type' => 'default',
+//				'target' => '',
+//				'size' => ''
+//			),
+//			$atts
+//		);
+//
+//		$type = (!empty($args['type'])) ? ' btn-' . $args['type'] : '';
+//		$link = $args['link'];
+//		$target = $args['target'];
+//		$size = (!empty($args['size'])) ? ' btn-' . $args['size'] : '';
+//
+//		$output = '<a class="btn ' . $type . $size . '" href="' . $link . '" target="' . $target . '"><span>' . \do_shortcode($content) . '</span></a>';
+//
+//		return $output;
+//	} // END public function shortcodeButton($atts, $content = null)
 
 	public function shortcodeCredits($atts, $content = null) {
 		$attributes = \shortcode_atts(array(